Add support for consts in jsdoc
