+++ scripts/jsdoc-automation/src/AIService/index.ts
@@ -1,0 +1,2 @@
+export * from './AIService.js';
+export * from './types';
