+++ scripts/jsdoc-automation/src/AIService/types/index.ts
@@ -1,0 +1,18 @@
+import { ASTQueueItem } from "../../types";
+
+export interface FileDocsGroup {
+    filePath: string;
+    classes: ASTQueueItem[];
+    methods: ASTQueueItem[];
+    interfaces: ASTQueueItem[];
+    types: ASTQueueItem[];
+    functions: ASTQueueItem[];
+}
+
+export interface OrganizedDocs {
+    classes: ASTQueueItem[];
+    methods: ASTQueueItem[];
+    interfaces: ASTQueueItem[];
+    types: ASTQueueItem[];
+    functions: ASTQueueItem[];
+}
