+++ scripts/jsdoc-automation/src/JSDocValidator.ts
@@ -1,5 +1,5 @@
 import { parse, ParserOptions } from "@typescript-eslint/parser";
-import { AIService } from "./AIService.js";
+import { AIService } from "./AIService/AIService.js";
 
 export class JSDocValidator {
     private parserOptions: ParserOptions = {
