+++ scripts/jsdoc-automation/src/JsDocGenerator.ts
@@ -1,4 +1,4 @@
-import { AIService } from "./AIService.js";
+import { AIService } from "./AIService/AIService.js";
 import { ASTQueueItem } from "./types/index.js";
 
 /**
@@ -50,7 +50,7 @@ export class JsDocGenerator {
         \`\`\`typescript
         ${queueItem.code}
         \`\`\`
-        
+
         Only return the JSDoc comment, not the code itself.
         `;
     }
@@ -61,7 +61,7 @@ export class JsDocGenerator {
         Class name: ${queueItem.code.match(/class (\w+)/)?.[1]}
 
         Only return the JSDoc for the Class itself, not the methods or anything in the class.
-        
+
         Only return the JSDoc comment for the class, no other text or code.
 
         Example:
