+++ scripts/jsdoc-automation/src/types/index.ts
@@ -1,7 +1,5 @@
 import { TSESTree } from "@typescript-eslint/types";
 
-import { TSESTree } from "@typescript-eslint/types";
-
 export interface ASTQueueItem {
     name: string;
     filePath: string;
@@ -30,14 +28,6 @@ export interface PrModeFileChange extends FullModeFileChange {
     deletions: number;
     changes: number;
     contents_url: string;
-}
-
-export interface OrganizedDocs {
-    classes: ASTQueueItem[];
-    methods: ASTQueueItem[];
-    interfaces: ASTQueueItem[];
-    types: ASTQueueItem[];
-    functions: ASTQueueItem[];
 }
 
 export interface TodoSection {
